% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge-kg.R
\name{.ensureHub}
\alias{.ensureHub}
\title{Check and load PhysioAnnotationHub}
\usage{
.ensureHub(hub)
}
\arguments{
\item{hub}{An existing hub object or NULL.}
}
\value{
A PhysioAnnotationHub object.
}
\description{
Check and load PhysioAnnotationHub
}
\keyword{internal}
