% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge-clinical.R
\name{.ensureHypergraph}
\alias{.ensureHypergraph}
\title{Ensure an MSKHypergraph is available}
\usage{
.ensureHypergraph(hg)
}
\arguments{
\item{hg}{An MSKHypergraph or NULL.}
}
\value{
An MSKHypergraph object.
}
\description{
Ensure an MSKHypergraph is available
}
\keyword{internal}
