% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msknet-hypergraph.R
\name{hyperedgeDegree}
\alias{hyperedgeDegree}
\title{Hyperedge (Muscle) Degree}
\usage{
hyperedgeDegree(hg)
}
\arguments{
\item{hg}{An MSKHypergraph object or incidence matrix.}
}
\value{
Named numeric vector of muscle degrees.
}
\description{
Returns the degree of each hyperedge (muscle), i.e., the number of
bones each muscle attaches to.
}
\examples{
C <- matrix(c(1, 1, 0, 0,  0, 1, 1, 0,  0, 0, 1, 1,  1, 0, 0, 1),
            nrow = 4, dimnames = list(paste0("b", 1:4), paste0("m", 1:4)))
hyperedgeDegree(MSKHypergraph(C))
}
