% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msknet-metrics.R
\name{mskCloseness}
\alias{mskCloseness}
\title{Closeness Centrality}
\usage{
mskCloseness(hg, type = c("bone", "muscle"))
}
\arguments{
\item{hg}{An MSKHypergraph object.}

\item{type}{Character, "bone" or "muscle" projection.}
}
\value{
Named numeric vector of closeness centrality values.
}
\description{
Closeness Centrality
}
\examples{
C <- matrix(c(1, 1, 0, 0,  0, 1, 1, 0,  0, 0, 1, 1,  1, 0, 0, 1),
            nrow = 4, dimnames = list(paste0("b", 1:4), paste0("m", 1:4)))
mskCloseness(MSKHypergraph(C), type = "bone")
}
