% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msknet-metrics.R
\name{mskNetworkMetrics}
\alias{mskNetworkMetrics}
\title{MSK Network Metrics}
\usage{
mskNetworkMetrics(hg, type = c("bone", "muscle"))
}
\arguments{
\item{hg}{An MSKHypergraph object.}

\item{type}{Character, "bone" or "muscle" projection.}
}
\value{
A list with: degree, strength, clustering_coef, density
}
\description{
Computes standard graph metrics for the projected bone or muscle graph.
}
\examples{
C <- matrix(c(1, 1, 0, 0,  0, 1, 1, 0,  0, 0, 1, 1,  1, 0, 0, 1),
            nrow = 4, dimnames = list(paste0("b", 1:4), paste0("m", 1:4)))
m <- mskNetworkMetrics(MSKHypergraph(C), type = "bone")
m$density
m$degree
}
