% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis-network.R
\name{plotCommunityStructure}
\alias{plotCommunityStructure}
\title{Plot Community Structure}
\usage{
plotCommunityStructure(community, ...)
}
\arguments{
\item{community}{Result from mskCommunityDetect or mskConsensusPartition.}

\item{...}{Additional arguments.}
}
\value{
Invisible NULL.
}
\description{
Visualizes detected communities with size information.
}
\examples{
cm <- mskCommunityDetect(MSKHypergraph(), gamma = 4.3)
plotCommunityStructure(cm)
}
