% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msknet-hypergraph.R
\name{projectBoneGraph}
\alias{projectBoneGraph}
\title{Project to Bone-centric Graph}
\usage{
projectBoneGraph(hg)
}
\arguments{
\item{hg}{An MSKHypergraph object, or a sparse incidence matrix.}
}
\value{
A symmetric sparse Matrix (n_bones x n_bones)
}
\description{
Creates the bone-bone weighted adjacency matrix A = t(C) \%*\% C where
A[i,j] counts the number of muscles shared between bones i and j.
This is the one-mode projection onto the bone (vertex) space.
}
\examples{
C <- matrix(c(1, 1, 0, 0,  0, 1, 1, 0,  0, 0, 1, 1,  1, 0, 0, 1),
            nrow = 4, dimnames = list(paste0("b", 1:4), paste0("m", 1:4)))
hg <- MSKHypergraph(C)
as.matrix(projectBoneGraph(hg))
}
\references{
Murphy AC et al. (2018) PLOS Biology.
}
