% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msknet-hypergraph.R
\name{projectMuscleGraph}
\alias{projectMuscleGraph}
\title{Project to Muscle-centric Graph}
\usage{
projectMuscleGraph(hg)
}
\arguments{
\item{hg}{An MSKHypergraph object, or a sparse incidence matrix.}
}
\value{
A symmetric sparse Matrix (n_muscles x n_muscles)
}
\description{
Creates the muscle-muscle weighted adjacency matrix B = C \%*\% t(C) where
B[i,j] counts the number of bones shared between muscles i and j.
This is the one-mode projection onto the muscle (hyperedge) space.
}
\examples{
C <- matrix(c(1, 1, 0, 0,  0, 1, 1, 0,  0, 0, 1, 1,  1, 0, 0, 1),
            nrow = 4, dimnames = list(paste0("b", 1:4), paste0("m", 1:4)))
hg <- MSKHypergraph(C)
as.matrix(projectMuscleGraph(hg))
}
\references{
Murphy AC et al. (2018) PLOS Biology.
}
