% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hill-model.R
\name{hillMuscleForce}
\alias{hillMuscleForce}
\title{Hill-type muscle force}
\usage{
hillMuscleForce(
  activation,
  norm_len,
  norm_vel,
  max_isometric_force = 1000,
  pennation = 0
)
}
\arguments{
\item{activation}{Activation in \verb{[0, 1]}.}

\item{norm_len}{Normalised fiber length.}

\item{norm_vel}{Normalised fiber velocity.}

\item{max_isometric_force}{Peak isometric force (N).}

\item{pennation}{Pennation angle (rad).}
}
\value{
Muscle force in N.
}
\description{
Total muscle-tendon force from activation and normalised fiber kinematics:
\eqn{F = F_{max}\,[a\,f_L(\tilde l)\,f_V(\tilde v) + f_{PE}(\tilde l)]\cos\alpha}.
}
