% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msknet-metrics.R
\name{mskNetworkMetrics}
\alias{mskNetworkMetrics}
\title{MSK Network Metrics}
\usage{
mskNetworkMetrics(hg, type = c("bone", "muscle"))
}
\arguments{
\item{hg}{An MSKHypergraph object.}

\item{type}{Character, "bone" or "muscle" projection.}
}
\value{
A list with: degree, strength, clustering_coef, density
}
\description{
Computes standard graph metrics for the projected bone or muscle graph.
}
