% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msknet-nullmodel.R
\name{mskNullEnsemble}
\alias{mskNullEnsemble}
\title{Generate Null Hypergraph Ensemble}
\usage{
mskNullEnsemble(hg = NULL, n_null = 100L, sim_params = list(), verbose = TRUE)
}
\arguments{
\item{hg}{An MSKHypergraph object.}

\item{n_null}{Number of null models to generate (default: 100).}

\item{sim_params}{List of simulation parameters (dt, n_steps, beta).}

\item{verbose}{Logical, print progress.}
}
\value{
A list with:
\describe{
\item{null_scores}{Matrix of impact scores (n_muscles x n_null)}
\item{null_degree_scores}{List, impact scores grouped by degree for each null}
\item{n_null}{Number of null models}
}
}
\description{
Creates multiple null hypergraphs and computes impact scores for each.
This is used to compute impact deviations relative to the null distribution.
}
\references{
Murphy AC et al. (2018) PLOS Biology.
}
