% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msknet-metrics.R
\name{mskShortestPaths}
\alias{mskShortestPaths}
\title{Shortest Path Distances}
\usage{
mskShortestPaths(hg, type = c("bone", "muscle"))
}
\arguments{
\item{hg}{An MSKHypergraph object.}

\item{type}{Character, "bone" or "muscle" projection.}
}
\value{
A numeric matrix of shortest path distances.
}
\description{
Computes shortest path distance matrix for the projected graph.
}
