% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msknet-hypergraph.R
\name{projectMuscleGraph}
\alias{projectMuscleGraph}
\title{Project to Muscle-centric Graph}
\usage{
projectMuscleGraph(hg)
}
\arguments{
\item{hg}{An MSKHypergraph object, or a sparse incidence matrix.}
}
\value{
A symmetric sparse Matrix (n_muscles x n_muscles)
}
\description{
Creates the muscle-muscle weighted adjacency matrix B = C \%*\% t(C) where
B[i,j] counts the number of bones shared between muscles i and j.
This is the one-mode projection onto the muscle (hyperedge) space.
}
\references{
Murphy AC et al. (2018) PLOS Biology.
}
