% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msknet-hypergraph.R
\name{vertexDegree}
\alias{vertexDegree}
\title{Vertex (Bone) Degree}
\usage{
vertexDegree(hg)
}
\arguments{
\item{hg}{An MSKHypergraph object or incidence matrix.}
}
\value{
Named numeric vector of bone degrees.
}
\description{
Returns the degree of each vertex (bone), i.e., the number of
muscles attached to each bone.
}
